Refactor the database classes to fight bloat.
<h1>Rationale:</h1>

This commit morphs the two classes, DatabaseManager and DatabaseListManager into a half dozen or so smaller and more focused classes: one each for groups, rooms, messages, members, and joins.  No functional changes are included.
