+++ app/src/main/java/com/pajato/android/gamechat/chat/ChatManager.java
@@ -23,12 +23,12 @@ import android.support.v4.app.FragmentManager;
 import android.util.Log;
 
 import com.pajato.android.gamechat.R;
-import com.pajato.android.gamechat.account.AccountManager;
+import com.pajato.android.gamechat.database.AccountManager;
 import com.pajato.android.gamechat.chat.adapter.ChatListItem;
 import com.pajato.android.gamechat.chat.model.Message;
 import com.pajato.android.gamechat.common.BaseFragment;
 import com.pajato.android.gamechat.common.Dispatcher;
-import com.pajato.android.gamechat.database.DatabaseListManager;
+import com.pajato.android.gamechat.database.MessageManager;
 import com.pajato.android.gamechat.main.NetworkManager;
 
 import java.util.HashMap;
@@ -85,7 +85,7 @@ public enum ChatManager {
         // In each case, return an empty dispatcher but for the fragment type of the next screen to
         // show.
         Map<String, Map<String, Map<String, Message>>> messageMap;
-        messageMap = DatabaseListManager.instance.messageMap;
+        messageMap = MessageManager.instance.messageMap;
         if (!NetworkManager.instance.isConnected()) return new Dispatcher<>(offline);
         if (!AccountManager.instance.hasAccount()) return new Dispatcher<>(signedOut);
         if (messageMap.size() == 0) return new Dispatcher<>(noMessages);
