+++ app/src/main/res/values/integers.xml
@@ -6,5 +6,4 @@
     <!-- Chat list type constants -->
     <integer name="groupList">0</integer>
     <integer name="messagesList">1</integer>
-    <integer name="roomList">2</integer>
 </resources>
